requirements.txt
openai
pinecone
sentence_transformers
python-dotenv
langchain-text-splitters
langchain-community
langgraph
langchain-openai
langchain-huggingface
langchain-pinecone